18 Основной функционал кабинета - исправлены пути к статическим ресурсам - Производитель, Модель и ККМ перенесены в схему equip - добавлен ОФД

+++ src/main/java/ru/antelit/fiskabinet/service/KkmService.java
@@ -25,7 +25,7 @@ public class KkmService {
         return repository.getKkmByTradepoint(tradepoint);
     }
 
-    public Kkm get(Long id) {
+    public Kkm get(Integer id) {
         return repository.getReferenceById(id);
     }
